requirements.txt
pandas==2.2.3
prophet==1.1.4
google-cloud-bigquery==3.14.1
scikit-learn==1.3.2
matplotlib==3.9.2
plotly==5.18.0
holidays==0.38
requests>=2.28.0
openmeteo-requests
requests-cache
